Move smallrye-common-config packages to io.smallrye.config.common
