+++ common/src/main/java/io/smallrye/config/common/utils/StringUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.smallrye.config.utils;
+package io.smallrye.config.common.utils;
 
 import java.util.ArrayList;
 import java.util.regex.Matcher;
